Fixes to the spec.
These are mostly moving fields around, specifically "Special"
fields, which are often in the "Possible Values" column.

On a few occasions, there were constraints both in the
"Possible Value" column and the "Special" column, in which case
I combined them with "&&".  I am not certain that this is correct
because I don't fully understand some of the extralogical functions
(e.g., Ignore, IsMeaningful, etc.)