Add ISO 19593-1 Processing Steps
Note that detailed data integrity rules are not implemented because ISO 19593-1 allows custom names (but only 2nd class!) so there is no point (i.e. adding "*" to PossibleValues). Custom names are not indicated in the model so that validators can warn when custom names are used since this indicates potential interoperability issues. Double underscore names with GWG_ prefix are CORRECT!!